Zainabdev sprint2

+++ src/main/java/com/uol/smqa/service/EventService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.uol.smqa.model.Event;
 import com.uol.smqa.model.EventType;
 import com.uol.smqa.repository.EventRepository;
+import com.uol.smqa.repository.OrganizerRepository;
 
 import jakarta.transaction.Transactional;
 
@@ -49,6 +50,9 @@ public class EventService {
     private EventRepository eventRepository;
     
     @Autowired
+    private OrganizerRepository organizerRepository;
+    
+    @Autowired
     private CustomerRepository customerRepository;
     @Autowired
     private EventTypeService eventTypeService;
@@ -76,6 +80,10 @@ public class EventService {
 	}
 
     public Event createEvent(Event event) {
+        return eventRepository.save(event);
+    }
+    
+    public Event createEventByAdmin(Event event) {
         return eventRepository.save(event);
     }
 
@@ -131,8 +139,14 @@ public class EventService {
 		Event event = this.eventRepository.findById(eventId);
 
 		if (event != null) 
-		{	this.eventRepository.delete(event);
+		{	
+			if(event.getBookedCustomers().size()==0)
+			{this.eventRepository.delete(event);
 		return "Event: "+eventId+" deleted Successfully!";}
+			else {
+				return "Event can't be deleted as customers have booked it";
+			}
+		}
 		else {
 			return "Event not found!";
 		}
